add i18n config, proxy and @formatjs/intl-localematcher
