+++ app/[locale]/page.tsx
@@ -1,0 +1,23 @@
+import Banner from "@/components/Banner";
+import About from "@/components/About";
+import Skills from "@/components/Skills";
+import Projects from "@/components/Projects";
+import Contact from "@/components/Contact";
+import TopButton from "@/components/TopButton";
+import Footer from "@/components/Footer";
+
+export default function LocaleHomePage() {
+  return (
+    <div className="min-h-screen w-full overflow-x-hidden">
+      <Banner />
+      <main>
+        <About />
+        <Skills />
+        <Projects />
+        <Contact />
+      </main>
+      <Footer />
+      <TopButton />
+    </div>
+  );
+}
