+++ app/page.tsx
@@ -1,23 +1,6 @@
-import Banner from "@/components/Banner";
-import About from "@/components/About";
-import Skills from "@/components/Skills";
-import Projects from "@/components/Projects";
-import Contact from "@/components/Contact";
-import TopButton from "@/components/TopButton";
-import Footer from "@/components/Footer";
+import { redirect } from "next/navigation";
+import { i18n } from "@/i18n.config";
 
 export default function Home() {
-  return (
-    <div className="min-h-screen w-full overflow-x-hidden">
-      <Banner />
-      <main>
-        <About />
-        <Skills />
-        <Projects />
-        <Contact />
-      </main>
-      <Footer />
-      <TopButton />
-    </div>
-  );
+  redirect(`/${i18n.defaultLocale}`);
 }
