+++ dictionaries/dictionary.en.ts
@@ -1,0 +1,62 @@
+import type { Dictionary } from "./types";
+
+export const dictionaryEn: Dictionary = {
+  languageNames: {
+    en: "English",
+    fr: "French",
+    ja: "Japanese",
+  },
+  nav: {
+    about: "About",
+    skills: "Skills",
+    projects: "Projects",
+    contact: "Contact",
+    languageLabel: "Language",
+  },
+  banner: {
+    intro: "Hi, my name is",
+    role: "I am a Full Stack Developer.",
+    description:
+      "I build responsive, production-ready web applications using TypeScript, React, and Next.js. I focus on creating intuitive user interfaces backed by reliable systems, translating complex ideas into seamless, maintainable experiences.",
+    cta: "View Work",
+  },
+  about: {
+    title: "About Me",
+    heading: "Hi! I am Colin, nice to meet you. Please take a look around.",
+    paragraphOne:
+      "I am a frontend-focused full-stack developer based in Yokohama, Japan, passionate about building user-centric web applications. Over the past few years, I have developed and maintained production-ready systems using TypeScript, React, and Next.js, with a strong emphasis on performance, reliability, and clean architecture.",
+    paragraphTwo:
+      "I enjoy translating complex ideas into intuitive, responsive interfaces, and I thrive in collaborative environments where thoughtful design and solid engineering come together to deliver real value.",
+    connect: "Let's connect.",
+  },
+  sections: {
+    skills: "Skills",
+    projects: "Projects",
+    contact: "Contact",
+  },
+  skills: {
+    frontend: "Front-end",
+    backend: "Back-end",
+    tools: "Tools",
+  },
+  projectCard: {
+    stack: "Stack",
+    onGithub: "on GitHub",
+    visitProject: "Visit",
+    liveSite: "Live site",
+  },
+  contact: {
+    intro: "Have a project in mind? Send me a message and I will get back to you.",
+    nameLabel: "Name",
+    emailLabel: "Email",
+    messageLabel: "Message",
+    sending: "Sending...",
+    submit: "Let's Collaborate",
+    successToast: "Message received. Thank you.",
+    unknownError: "Something went wrong. Please try again.",
+    networkError: "Network error. Please try again.",
+  },
+  footer: {
+    copyright: "© Colin Bertin JP 2026, All rights reserved.",
+  },
+};
