+++ dictionaries/dictionary.fr.ts
@@ -1,0 +1,63 @@
+import type { Dictionary } from "./types";
+
+export const dictionaryFr: Dictionary = {
+  languageNames: {
+    en: "Anglais",
+    fr: "Français",
+    ja: "Japonais",
+  },
+  nav: {
+    about: "À propos",
+    skills: "Compétences",
+    projects: "Projets",
+    contact: "Contact",
+    languageLabel: "Langue",
+  },
+  banner: {
+    intro: "Bonjour, je m'appelle",
+    role: "Je suis développeur full stack.",
+    description:
+      "Je construis des applications web réactives et prêtes pour la production avec TypeScript, React et Next.js. Je crée des interfaces intuitives appuyées par des systèmes fiables, en transformant des idées complexes en expériences fluides et maintenables.",
+    cta: "Voir les projets",
+  },
+  about: {
+    title: "À propos",
+    heading: "Salut ! Je suis Colin, ravi de vous rencontrer. Bonne visite.",
+    paragraphOne:
+      "Je suis un développeur full stack orienté front-end basé à Yokohama, au Japon, passionné par la création d'applications web centrées sur l'utilisateur. Ces dernières années, j'ai développé et maintenu des systèmes prêts pour la production avec TypeScript, React et Next.js, avec une attention particulière à la performance, la fiabilité et une architecture propre.",
+    paragraphTwo:
+      "J'aime transformer des idées complexes en interfaces intuitives et réactives, et je m'épanouis dans des environnements collaboratifs où design réfléchi et ingénierie solide se rejoignent pour apporter une vraie valeur.",
+    connect: "Discutons ensemble.",
+  },
+  sections: {
+    skills: "Compétences",
+    projects: "Projets",
+    contact: "Contact",
+  },
+  skills: {
+    frontend: "Front-end",
+    backend: "Back-end",
+    tools: "Outils",
+  },
+  projectCard: {
+    stack: "Stack",
+    onGithub: "sur GitHub",
+    visitProject: "Visiter",
+    liveSite: "Site en ligne",
+  },
+  contact: {
+    intro:
+      "Vous avez un projet en tête? Envoyez-moi un message et je vous répondrai rapidement.",
+    nameLabel: "Nom",
+    emailLabel: "Email",
+    messageLabel: "Message",
+    sending: "Envoi...",
+    submit: "Collaborons ensemble",
+    successToast: "Message reçu. Merci.",
+    unknownError: "Une erreur est survenue. Veuillez réessayer.",
+    networkError: "Erreur réseau. Veuillez réessayer.",
+  },
+  footer: {
+    copyright: "© Colin Bertin JP 2026, Tous droits réservés.",
+  },
+};
