+++ dictionaries/dictionary.ja.ts
@@ -1,0 +1,62 @@
+import type { Dictionary } from "./types";
+
+export const dictionaryJa: Dictionary = {
+  languageNames: {
+    en: "英語",
+    fr: "フランス語",
+    ja: "日本語",
+  },
+  nav: {
+    about: "自己紹介",
+    skills: "スキル",
+    projects: "プロジェクト",
+    contact: "お問い合わせ",
+    languageLabel: "言語",
+  },
+  banner: {
+    intro: "こんにちは、私は",
+    role: "フルスタック開発者です。",
+    description:
+      "TypeScript、React、Next.jsを使って、レスポンシブで本番運用可能なウェブアプリケーションを開発しています。複雑なアイデアを、保守しやすく直感的な体験に落とし込むことを大切にしています。",
+    cta: "実績を見る",
+  },
+  about: {
+    title: "自己紹介",
+    heading: "はじめまして、Colinです。ぜひご覧ください。",
+    paragraphOne:
+      "横浜を拠点に活動する、フロントエンド志向のフルスタック開発者です。ここ数年はTypeScript、React、Next.jsを中心に、本番運用向けシステムの開発と保守を行ってきました。性能、信頼性、そしてクリーンな設計を重視しています。",
+    paragraphTwo:
+      "複雑な課題を直感的で使いやすいUIに変換することが得意です。丁寧なデザインと堅実な実装が合わさるチーム環境で、価値あるプロダクトを届けることにやりがいを感じています。",
+    connect: "ぜひご連絡ください。",
+  },
+  sections: {
+    skills: "スキル",
+    projects: "プロジェクト",
+    contact: "お問い合わせ",
+  },
+  skills: {
+    frontend: "フロントエンド",
+    backend: "バックエンド",
+    tools: "ツール",
+  },
+  projectCard: {
+    stack: "技術スタック",
+    onGithub: "を GitHub で見る",
+    visitProject: "アクセス",
+    liveSite: "公開サイト",
+  },
+  contact: {
+    intro: "プロジェクトのご相談があれば、メッセージをお送りください。",
+    nameLabel: "お名前",
+    emailLabel: "メール",
+    messageLabel: "メッセージ",
+    sending: "送信中...",
+    submit: "相談する",
+    successToast: "メッセージを受け取りました。ありがとうございます。",
+    unknownError: "エラーが発生しました。もう一度お試しください。",
+    networkError: "ネットワークエラーです。もう一度お試しください。",
+  },
+  footer: {
+    copyright: "© Colin Bertin JP 2026, All rights reserved.",
+  },
+};
