+++ i18n.config.ts
@@ -1,0 +1,6 @@
+export const i18n = {
+  defaultLocale: "en",
+  locales: ["en", "ja", "fr"],
+} as const;
+
+export type Locale = (typeof i18n)["locales"][number];
