+++ package.json
@@ -9,6 +9,7 @@
     "lint": "eslint"
   },
   "dependencies": {
+    "@formatjs/intl-localematcher": "^0.8.2",
     "@react-email/render": "^2.0.4",
     "devicon": "^2.17.0",
     "next": "16.1.4",
